feat(tt): move logic to machine
